app/requirements.txt
Flask==2.0.1
Flask-MySQLdb==0.2.0
requests==2.26.0
Werkzeug==2.2.2
python-dotenv==0.19.2
Flask-Cors==3.0.10
prometheus_client==0.17.1 
prometheus_flask_exporter==0.22.3
gunicorn==20.1.0
